M    trunk/include/types.h A    trunk/include/libcompetitivefunctions.h M    trunk/include/liboutputfunctions.h M    trunk/include/backpropagationalgo.h A    trunk/include/libperiodicfunctions.h M    trunk/include/realvec.h M    trunk/src/nnfwfactory.cpp A    trunk/src/libcompetitivefunctions.cpp A    trunk/src/libperiodicfunctions.cpp M    trunk/src/liboutputfunctions.cpp M    trunk/src/neuralnet.cpp M    trunk/CHANGELOG M    trunk/CMakeLists.txt A    flexynet/include/propertizedmodel.h A    flexynet/Doxyfile A    flexynet/src/propertizedmodel.cpp

+++ trunk/src/neuralnet.cpp
@@ -368,7 +368,7 @@ BaseNeuralNet* BaseNeuralNet::clone() const {
 	}
 	// --- copy the order -- not-efficient
 	UpdatableVec ord;
-	for( int i=0; i<order().size(); i++ ) {
+	for( int i=0; i<(int)order().size(); i++ ) {
 		ord << clone->getByName( order()[i]->name() );
 	}
 	clone->setOrder( ord );
